+++ src/main/java/com/kinshelf/controllers/BookController.java
@@ -28,20 +28,20 @@ public class BookController {
     }
     
     @GetMapping("/{id}")
-    public ResponseEntity<BookResponseDTO> getById(@PathVariable Integer id) {
+    public ResponseEntity<BookResponseDTO> getById(@PathVariable Long id) {
         return ResponseEntity.ok(bookService.findById(id));
     }
     
     @PutMapping("/{id}")
     public ResponseEntity<BookResponseDTO> update(
-            @PathVariable Integer id,
+            @PathVariable Long id,
             @Valid @RequestBody BookCreateDTO dto
     ) {
         return ResponseEntity.ok(bookService.update(id, dto));
     }
     
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> delete(@PathVariable Integer id) {
+    public ResponseEntity<Void> delete(@PathVariable Long id) {
         bookService.delete(id);
         return ResponseEntity.noContent().build();
     }
